notorch_requirements.txt
click==8.0.3
matplotlib==3.5.0
numpy==1.19.2
python-dotenv==0.19.2
scikit_learn==1.0.2
setuptools==58.0.4